backend/requirements.txt
claude-agent-sdk
openai-agents
python-dotenv
fastapi
uvicorn[standard]
pydantic